+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 ## [Unreleased]
+### Added
+- feat: Stop at height full-node arguments
+  **New hidden node args:**\
+  `--stop-at-l1-height`: Stop the full node L1 sync after reaching this L1 height
+  `--stop-at-l2-height`: Stop the full node L2 sync after reaching this L2 height
 
 ## [v2.2.0](2026-03-02)
 ### Added